+++ api/python/src/PDB/pyCompilationUnit.cpp
@@ -1,6 +1,7 @@
 #include <sstream>
 #include "LIEF/PDB/CompilationUnit.hpp"
 #include "LIEF/PDB/BuildMetadata.hpp"
+#include "LIEF/DebugDeclOpt.hpp"
 #include "PDB/pyPDB.hpp"
 
 #include <nanobind/stl/string.h>
@@ -57,6 +58,11 @@ void create<pdb::CompilationUnit>(nb::module_& m) {
     )
     .def_prop_ro("build_metadata", &CompilationUnit::build_metadata,
                  nb::rv_policy::take_ownership, nb::keep_alive<0, 1>())
+
+    .def("to_decl", &pdb::CompilationUnit::to_decl,
+         "Generates a C/C++ definition for the functions defined in this "
+         "compilation unit"_doc,
+         "opt"_a = DeclOpt())
   LIEF_DEFAULT_STR(CompilationUnit);
 }