+++ api/rust/crates/lief-ffi/src/debug_decl_opt.rs
@@ -11,6 +11,7 @@ pub mod ffi {
         fn set_is_cpp(self: Pin<&mut LIEF_DeclOpt>, value: bool);
         fn set_show_extended_annotations(self: Pin<&mut LIEF_DeclOpt>, value: bool);
         fn set_include_types(self: Pin<&mut LIEF_DeclOpt>, value: bool);
+        fn set_include_locals(self: Pin<&mut LIEF_DeclOpt>, value: bool);
         fn set_desugar(self: Pin<&mut LIEF_DeclOpt>, value: bool);
     }
     impl UniquePtr<LIEF_DeclOpt> {}