+++ api/rust/crates/lief-ffi/src/pdb/function.rs
@@ -4,6 +4,7 @@ pub mod ffi {
         include!("LIEF/rust/PDB/Function.hpp");
 
         type DebugLocation = crate::debug_location::ffi::DebugLocation;
+        type LIEF_DeclOpt = crate::debug_decl_opt::ffi::LIEF_DeclOpt;
 
         type PDB_Function;
 
@@ -13,6 +14,8 @@ pub mod ffi {
         fn section_name(self: &PDB_Function) -> UniquePtr<CxxString>;
         fn debug_location(self: &PDB_Function) -> UniquePtr<DebugLocation>;
         fn to_string(self: &PDB_Function) -> UniquePtr<CxxString>;
+        fn to_decl(self: &PDB_Function) -> UniquePtr<CxxString>;
+        fn to_decl_with_opt(self: &PDB_Function, opt: &LIEF_DeclOpt) -> UniquePtr<CxxString>;
     }
 
     impl UniquePtr<PDB_Function> {}