+++ api/rust/include/LIEF/rust/PDB/CompilationUnit.hpp
@@ -19,6 +19,7 @@
 #include "LIEF/rust/PDB/Function.hpp"
 #include "LIEF/rust/PDB/BuildMetadata.hpp"
 #include "LIEF/rust/helpers.hpp"
+#include "LIEF/rust/DebugDeclOpt.hpp"
 #include "LIEF/PDB/CompilationUnit.hpp"
 
 class PDB_CompilationUnit : private Mirror<LIEF::pdb::CompilationUnit> {
@@ -74,6 +75,14 @@ class PDB_CompilationUnit : private Mirror<LIEF::pdb::CompilationUnit> {
 
   auto to_string() const {
     return to_unique_string(get().to_string());
+  }
+
+  auto to_decl() const {
+    return to_unique_string(get().to_decl());
+  }
+
+  auto to_decl_with_opt(const LIEF_DeclOpt& opt) const {
+    return to_unique_string(get().to_decl(opt.conf()));
   }
 };